# app/src/services/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/api/token/": {
      "post": {
        "tags": ["token"],
        "summary": "Open Id Login",
        "operationId": "open_id_login_api_token__post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TokenRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/Token" }, { "type": "null" }],
                  "title": "Response Open Id Login Api Token  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "tags": ["users"],
        "summary": "Read User",
        "operationId": "read_user_api_users__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserSchema" } }
            }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/notes/": {
      "get": {
        "tags": ["notes"],
        "summary": "Read Notes",
        "operationId": "read_notes_api_notes__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/NoteSchema" },
                  "type": "array",
                  "title": "Response Read Notes Api Notes  Get"
                }
              }
            }
          }
        },
        "security": [{ "Bearer": [] }]
      },
      "post": {
        "tags": ["notes"],
        "summary": "Create Note",
        "operationId": "create_note_api_notes__post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/NoteCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NoteSchema" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/notes/{note_id}": {
      "patch": {
        "tags": ["notes"],
        "summary": "Update Note",
        "operationId": "update_note_api_notes__note_id__patch",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Note Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/NoteUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NoteSchema" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NoteCreate": {
        "properties": { "title": { "type": "string", "title": "Title" } },
        "type": "object",
        "required": ["title"],
        "title": "NoteCreate"
      },
      "NoteSchema": {
        "properties": {
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "rich_text": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Rich Text"
          },
          "edited_at": {
            "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
            "title": "Edited At"
          },
          "user_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "User Id"
          }
        },
        "type": "object",
        "required": ["title", "rich_text"],
        "title": "NoteSchema"
      },
      "NoteUpdate": {
        "properties": {
          "title": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Title" },
          "rich_text": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Rich Text" }
        },
        "type": "object",
        "title": "NoteUpdate"
      },
      "Token": {
        "properties": {
          "access_token": { "type": "string", "title": "Access Token" },
          "token_type": { "type": "string", "title": "Token Type" }
        },
        "type": "object",
        "required": ["access_token", "token_type"],
        "title": "Token"
      },
      "TokenRequest": {
        "properties": { "id_token": { "type": "string", "title": "Id Token" } },
        "type": "object",
        "required": ["id_token"],
        "title": "TokenRequest"
      },
      "UserSchema": {
        "properties": {
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "email": { "type": "string", "title": "Email" },
          "given_name": { "type": "string", "title": "Given Name" },
          "family_name": { "type": "string", "title": "Family Name" }
        },
        "type": "object",
        "required": ["email", "given_name", "family_name"],
        "title": "UserSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT token",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}